Polish utils module
