+++ packages/toolkit/src/helpers/index.js
@@ -2,11 +2,9 @@ export * from './ipfs'
 export * from './node'
 export * from './wrapper'
 export * from './constants'
-export * from './solidity-extractor'
 export * from './useEnvironment'
 export * from './web3-fallback'
 export * from './loadConfigFiles'
 export * from './truffle-config'
-export * from './resolveAddressOrEnsDomain'
 export { default as defaultAPMName } from './default-apm'
 export { default as extractContractInfoToFile } from './extractContractInfoToFile'
