+++ packages/toolkit/src/index.js
@@ -3,3 +3,4 @@ export * from './dao'
 export * from './helpers'
 export * from './utils'
 export * from './util'
+export * from './token'
