+++ packages/toolkit/src/utils/index.js
@@ -1,5 +1,6 @@
-export * from './token'
-export * from './params'
+export * from './aclParams'
+export * from './solidityExtractor'
+export * from './resolvers'
 export { default as encodeActCall } from './encodeActCall'
 export { default as encodeInitPayload } from './encodeInitPayload'
 export { default as generateArtifact } from './generateArtifact'
